+++ NEWS.md
@@ -7,7 +7,7 @@ peacesciencer 0.2.0
 - Add Correlates of War intergovernmental organizations data.
 - Add function for adding Correlates of War intergovernmental organizations data.
 - Add Correlates of War trade data (state-year and dyad-year).
-- Add minimum distance data
+- Add minimum distance data and `add_minimum_distance()` function
 - Add citations data frame for assistance in properly citing things.
 - Add `create_statedays()` function.